requirements.txt
streamlit>=1.28.0
numpy==1.26.0
opencv-python-headless>=4.5.0
pandas>=2.0.0
pillow>=9.0.0
scikit-learn>=1.0.0
tensorflow>=2.10.0
tflite-runtime>=2.10.0